third_party/blink/web_tests/external/wpt/content-security-policy/gen/top.meta/worker-src-self/worker-import.https-expected.txt

This is a testharness.js-based test.
[FAIL] Content Security Policy: Expects blocked for worker-import to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
  assert_equals: One violation event should be fired expected 1 but got 0
[FAIL] Content Security Policy: Expects blocked for worker-import to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
  assert_equals: One violation event should be fired expected 1 but got 0
[FAIL] Content Security Policy: Expects blocked for worker-import to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
  assert_equals: One violation event should be fired expected 1 but got 0
[FAIL] Content Security Policy: Expects blocked for worker-import to same-https origin and swap-origin redirection from https context.: securitypolicyviolation
  assert_equals: One violation event should be fired expected 1 but got 0
Harness: the test ran to completion.
